Improve layer position updating
Got rid of the unique key index for dynamo_layer_position in the layers
table. Added a `move_layer` helper function in acp_dynamo.php. Fixed
a small lang constant issue with the dropdown menu. Added arrow icons
on the layers display page for moving layers around and changed the
edit and delete links to icons.