Refactor UCP module
Backend (fewer arrays and queries, and less duplication) and frontend
(added a "no item" option, separated layers into divs which makes for
nicer JS as well as a better no-js fallback, added real hyperlinks to
the layers-pane div, added some description info for the layers, made
some changes to the text and probably more).

For issue #42. The CSS still needs to be moved elsewhere, although
the necessity of that is debatable (especially if there's no other
way to have {IMAGE_HEIGHT} and {IMAGE_WIDTH} accessible).