Improve layer reordering
Improves upon the current method of moving a layer to a new position
and also adds a new way of moving layers around (up/down arrows). The
`position` unique index in the layers table is dropped. This may
cause problems when updating (if using a non-MySQL db) but hopefully
will not, and 99% of phpBB installations use MySQL anyways.

Closes #32, #33, #34.

Some relics of the log-adding spree that I forgot to attach to the
relevant commit are also included.